backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alpaca-trade-api>=3.2.0",
    "fastapi>=0.136.0",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "tavily-python>=0.7.23",
    "uvicorn>=0.44.0",
    "yfinance>=0.2.58",
]
